+++ lib/src/scrape/apps/lib/proc.py
@@ -1,4 +1,5 @@
 import multiprocessing
+import traceback
 
 
 class XtProcess(multiprocessing.Process):
