+++ lib/src/scrape/requirements.txt
@@ -1,2 +1,3 @@
 scrapy
 selenium
+fake-useragent
